Convert `Switch` global into a "namespace"
Closes #34.
