+++ packages/runtime/src/dns.ts
@@ -1,6 +1,15 @@
 import { $ } from './$';
 import { toPromise } from './utils';
 
-export function resolve(hostname: string) {
+/**
+ * Performs a DNS lookup to resolve a hostname to an array of IP addresses.
+ *
+ * @example
+ *
+ * ```typescript
+ * const ipAddresses = await Switch.resolveDns('example.com');
+ * ```
+ */
+export function resolveDns(hostname: string) {
 	return toPromise($.dnsResolve, hostname);
 }
