+++ packages/runtime/src/internal.ts
@@ -1,5 +1,5 @@
 import type { connect } from './tcp';
-import type { SocketOptions } from './types';
+import type { SocketOptions } from './switch';
 
 export const INTERNAL_SYMBOL = Symbol('Internal');
 
@@ -34,13 +34,3 @@ export type Keys = {
 	modifiers: bigint;
 	[i: number]: bigint;
 };
-
-export interface Vibration {
-	duration: number;
-	lowAmp: number;
-	lowFreq: number;
-	highAmp: number;
-	highFreq: number;
-}
-
-export type VibrationValues = Omit<Vibration, 'duration'>;
